Add getCollidables function
Adds the ability to get an array of all the collidable rectangles with their position relative to the map. This helps with people who are trying to compare location of a player to the collidables. For example, figuring out the side (left, right, top, bottom) that a player is colliding with.